Removed unnecessary for loop that once filled the salt.
Note that the "// NOOP" comment was misleading, the salt variable
actually was modified inside the for loop.

This functionality will never be used though as the salt is checked
some lines above to an exact length of 2. Filling it with "A" would
be a very bad idea anyway as e.g. glibc fills a one-char salt with
exactly the same char so our implementation would have been incompatible!