Support decoding of Base64 with intermediary padding.
Some base64 encoders in the wild incorrectly pad data before the end of
data. For instance, the string "testtesttest" might be encoded as
"dGVzdA==dGVzdA==dGVzdA==" instead of "dGVzdHRlc3R0ZXN0". This is,
technically speaking, a bug.

This patch makes the decoder tolerant to this kind of nonsense.
Intermediary padding is processed and then decoding carries on as
though a new string were provided.