[CODEC-265] BaseNCodec to expand buffer using overflow conscious code.
Add a test to encode a 1 Gigabyte byte[].

Added overflow conscious code for buffer expansion.

This is based on java.util.ArrayList but modified to use a local copy of
JDK 1.8 Integer.compareUnsigned(int, int).

Unlike the java.util.ArrayList if the minCapcity exceeds the head room
for the maximum array size, the capacity returned is the required
minCapcity and not Integer.MAX_VALUE.